split write_bounding specific to CSW3 impl, rename test cases to be consistent with the rest of the test suites, add GetRecordById fixes and tests
