+++ pycsw/plugins/profiles/ebrim/ebrim.py
@@ -31,6 +31,7 @@
 import os
 from pycsw.core.etree import etree
 from pycsw.core import config, util
+from pycsw.ogc.csw.csw2 import write_boundingbox
 from pycsw.plugins.profiles import profile
 
 class EBRIM(profile.Profile):
@@ -150,7 +151,7 @@ class EBRIM(profile.Profile):
             etree.SubElement(description, util.nspath_eval('rim:LocalizedString', self.namespaces), value=unicode(util.getqattr(result, self.context.md_core_model['mappings']['pycsw:Abstract'])))
 
             val = util.getqattr(result, self.context.md_core_model['mappings']['pycsw:BoundingBox'])
-            bboxel = util.write_boundingbox(val, self.context.namespaces)
+            bboxel = write_boundingbox(val, self.context.namespaces)
 
             if bboxel is not None:
                 bboxslot = etree.SubElement(node, util.nspath_eval('rim:Slot', self.namespaces),
