+++ tests/expected/suites_csw30_get_Exception-GetDomain-value-reference.xml
