+++ tests/expected/suites_csw30_post_Exception-GetDomain-valuereference-bad.xml
