+++ tests/expected/suites_csw30_post_Exception-bad-xml.xml
