+++ tests/suites/csw30/post/Exception-GetDomain-valuereference-bad.xml
