+++ tests/suites/csw30/post/Exception-bad-xml.xml
