+++ tests/suites/csw30/post/Exception-not-xml.xml
